requirements.txt
panoramisk
apscheduler
flask
gunicorn
requests
mysql-connector-python
celery
redis